refactor(app): ConnectVia
