+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -22,9 +22,9 @@ import { MaintenanceRunTakeover } from '../organisms/TakeoverModal'
 import { FirmwareUpdateTakeover } from '../organisms/FirmwareUpdateModal/FirmwareUpdateTakeover'
 import { IncompatibleModuleTakeover } from '../organisms/IncompatibleModule'
 import { EstopTakeover } from '../organisms/EmergencyStop'
-import { ConnectViaEthernet } from '../pages/ConnectViaEthernet'
-import { ConnectViaUSB } from '../pages/ConnectViaUSB'
-import { ConnectViaWifi } from '../pages/ConnectViaWifi'
+import { ConnectViaEthernet } from '../pages/ODD/ConnectViaEthernet'
+import { ConnectViaUSB } from '../pages/ODD/ConnectViaUSB'
+import { ConnectViaWifi } from '../pages/ODD/ConnectViaWifi'
 import { EmergencyStop } from '../pages/EmergencyStop'
 import { NameRobot } from '../pages/NameRobot'
 import { NetworkSetupMenu } from '../pages/NetworkSetupMenu'
