+++ app/src/pages/ODD/ConnectViaEthernet/__tests__/DisplayConnectionStatus.test.tsx
@@ -2,9 +2,9 @@ import * as React from 'react'
 import { vi, it, describe, expect, beforeEach } from 'vitest'
 import { fireEvent, screen } from '@testing-library/react'
 
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
-import { DisplayConnectionStatus } from '../../../pages/ConnectViaEthernet/DisplayConnectionStatus'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
+import { DisplayConnectionStatus } from '../DisplayConnectionStatus'
 
 import type { NavigateFunction } from 'react-router-dom'
