+++ app/src/pages/ODD/ConnectViaUSB/_tests__/ConnectedViaUSB.test.tsx
@@ -3,11 +3,11 @@ import { vi, it, describe, expect, beforeEach, afterEach } from 'vitest'
 import { MemoryRouter } from 'react-router-dom'
 import { fireEvent } from '@testing-library/react'
 
-import { renderWithProviders } from '../../../__testing-utils__'
+import { renderWithProviders } from '../../../../__testing-utils__'
 import { useConnectionsQuery } from '@opentrons/react-api-client'
 
-import { i18n } from '../../../i18n'
-import { ConnectViaUSB } from '../../../pages/ConnectViaUSB'
+import { i18n } from '../../../../i18n'
+import { ConnectViaUSB } from '../'
 
 import type { UseQueryResult } from 'react-query'
 import type { ActiveConnections } from '@opentrons/api-client'
