+++ app/src/pages/ODD/ConnectViaWifi/JoinOtherNetwork.tsx
@@ -3,10 +3,10 @@ import { useTranslation } from 'react-i18next'
 
 import { Flex, DIRECTION_COLUMN } from '@opentrons/components'
 
-import { SetWifiSsid } from '../../organisms/NetworkSettings'
-import { RobotSetupHeader } from '../../organisms/RobotSetupHeader'
+import { SetWifiSsid } from '../../../organisms/NetworkSettings'
+import { RobotSetupHeader } from '../../../organisms/RobotSetupHeader'
 
-import type { WifiScreenOption } from '../../pages/ConnectViaWifi'
+import type { WifiScreenOption } from './'
 
 interface JoinOtherNetworkProps {
   setCurrentOption: (option: WifiScreenOption) => void
