+++ app/src/pages/ODD/ConnectViaWifi/SelectAuthenticationType.tsx
@@ -3,11 +3,11 @@ import { useTranslation } from 'react-i18next'
 
 import { Flex, DIRECTION_COLUMN } from '@opentrons/components'
 
-import { SelectAuthenticationType as SelectAuthenticationTypeComponent } from '../../organisms/NetworkSettings'
-import { RobotSetupHeader } from '../../organisms/RobotSetupHeader'
+import { SelectAuthenticationType as SelectAuthenticationTypeComponent } from '../../../organisms/NetworkSettings'
+import { RobotSetupHeader } from '../../../organisms/RobotSetupHeader'
 
 import type { WifiSecurityType } from '@opentrons/api-client'
-import type { WifiScreenOption } from '../../pages/ConnectViaWifi'
+import type { WifiScreenOption } from './'
 
 interface SelectAuthenticationTypeProps {
   handleWifiConnect: () => void
