+++ app/src/pages/ODD/ConnectViaWifi/SetWifiCred.tsx
@@ -3,10 +3,10 @@ import { useTranslation } from 'react-i18next'
 
 import { Flex, DIRECTION_COLUMN } from '@opentrons/components'
 
-import { SetWifiCred as SetWifiCredComponent } from '../../organisms/NetworkSettings'
-import { RobotSetupHeader } from '../../organisms/RobotSetupHeader'
+import { SetWifiCred as SetWifiCredComponent } from '../../../organisms/NetworkSettings'
+import { RobotSetupHeader } from '../../../organisms/RobotSetupHeader'
 
-import type { WifiScreenOption } from '../../pages/ConnectViaWifi'
+import type { WifiScreenOption } from './'
 
 interface SetWifiCredProps {
   handleConnect: () => void
