+++ app/src/pages/ODD/ConnectViaWifi/__tests__/ConnectViaWifi.test.tsx
@@ -3,18 +3,18 @@ import { vi, it, describe, expect, beforeEach, afterEach } from 'vitest'
 import { MemoryRouter } from 'react-router-dom'
 import { fireEvent, screen } from '@testing-library/react'
 
-import { renderWithProviders } from '../../../__testing-utils__'
-import { i18n } from '../../../i18n'
-import * as RobotApi from '../../../redux/robot-api'
-import * as Fixtures from '../../../redux/networking/__fixtures__'
-import { useWifiList } from '../../../resources/networking/hooks'
-import * as Networking from '../../../redux/networking'
-import { ConnectViaWifi } from '../../../pages/ConnectViaWifi'
+import { renderWithProviders } from '../../../../__testing-utils__'
+import { i18n } from '../../../../i18n'
+import * as RobotApi from '../../../../redux/robot-api'
+import * as Fixtures from '../../../../redux/networking/__fixtures__'
+import { useWifiList } from '../../../../resources/networking/hooks'
+import * as Networking from '../../../../redux/networking'
+import { ConnectViaWifi } from '../'
 
-vi.mock('../../../redux/discovery')
-vi.mock('../../../resources/networking/hooks')
-vi.mock('../../../redux/networking/selectors')
-vi.mock('../../../redux/robot-api/selectors')
+vi.mock('../../../../redux/discovery')
+vi.mock('../../../../resources/networking/hooks')
+vi.mock('../../../../redux/networking/selectors')
+vi.mock('../../../../redux/robot-api/selectors')
 
 const mockWifiList = [
   { ...Fixtures.mockWifiNetwork, ssid: 'foo', active: true },
