+++ app/src/pages/ODD/ConnectViaWifi/index.tsx
@@ -4,19 +4,19 @@ import last from 'lodash/last'
 
 import { Flex, DIRECTION_COLUMN, SPACING } from '@opentrons/components'
 
-import { StepMeter } from '../../atoms/StepMeter'
-import { DisplayWifiList } from '../../organisms/NetworkSettings'
-import * as Networking from '../../redux/networking'
-import { getLocalRobot } from '../../redux/discovery'
-import * as RobotApi from '../../redux/robot-api'
-import { useWifiList } from '../../resources/networking/hooks'
-import { JoinOtherNetwork } from '../../pages/ConnectViaWifi/JoinOtherNetwork'
-import { SelectAuthenticationType } from '../../pages/ConnectViaWifi/SelectAuthenticationType'
-import { SetWifiCred } from '../../pages/ConnectViaWifi/SetWifiCred'
-import { WifiConnectStatus } from '../../pages/ConnectViaWifi/WifiConnectStatus'
+import { StepMeter } from '../../../atoms/StepMeter'
+import { DisplayWifiList } from '../../../organisms/NetworkSettings'
+import * as Networking from '../../../redux/networking'
+import { getLocalRobot } from '../../../redux/discovery'
+import * as RobotApi from '../../../redux/robot-api'
+import { useWifiList } from '../../../resources/networking/hooks'
+import { JoinOtherNetwork } from './JoinOtherNetwork'
+import { SelectAuthenticationType } from './SelectAuthenticationType'
+import { SetWifiCred } from './SetWifiCred'
+import { WifiConnectStatus } from './WifiConnectStatus'
 
 import type { WifiSecurityType } from '@opentrons/api-client'
-import type { State } from '../../redux/types'
+import type { State } from '../../../redux/types'
 
 const WIFI_LIST_POLL_MS = 5000
 export type WifiScreenOption =
